requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
frozenlist==1.3.3
idna==3.4
magic-filter==1.0.9
multidict==6.0.4
Pillow==10.0.0
pypng==0.20220715.0
python-dotenv==1.0.0
pytz==2023.3
qrcode==7.4.2
typing_extensions==4.7.1
yarl==1.9.2
requests==2.31.0
deep-translator==1.11.4
psycopg2-binary==2.9.9
